+++ lib/twenv/command_mixins/tweet_filter.rb
@@ -11,13 +11,13 @@ module TWEnv::Command::TweetFilter
     end
   end
 
-  def is_reply_to_filter!(tweets)
+  def is_reply_to_filter!(tweets, screen_name)
     tweets.select! do |tweet|
       next unless tweet.reply?
       user = client.user(tweet.in_reply_to_user_id)
-      user.screen_name == opts['is-reply-to']
+      user.screen_name == screen_name
     rescue Twitter::Error::NotFound, Twitter::Error::Forbidden
       nil
-    end if opts['is-reply-to']
+    end if screen_name
   end
 end